@@ -39,11 +39,12 @@
     "immutable-kv": "^1.2.0",
     "jquery": "^3.1.1",
     "notify-promise": "0.0.1",
-    "react": "^15.4.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
     "react-css-modules": "^3.7.10",
-    "react-dnd": "^2.3.0",
-    "react-dnd-html5-backend": "^2.3.0",
-    "react-dom": "^15.4.2",
+    "react-dnd": "^2.4.0",
+    "react-dnd-html5-backend": "^2.4.1",
+    "react-dom": "^15.5.4",
     "react-redux": "^4.4.5",
     "react-router": "^4.0.0-alpha.4",
     "redux": "^3.6.0",
