@@ -5,7 +5,7 @@ import { HISTORY_UNDO, HISTORY_REDO, HISTORY_CLEAN } from '../actions/history';
 const RECORD_ACTIONS = [
 	ROUTER_TO,
 	ROUTER_REDIRECT,
-	KV_MOVE, KV_TOGGLE, KV_SELECT, KV_SWITCH_TAB, KV_MODIFY, KV_CREATE, KV_CREATE_GROUP
+	KV_MOVE, KV_TOGGLE, KV_SELECT, KV_SWITCH_TAB, KV_MODIFY, KV_CREATE, KV_CREATE_GROUP,
 ];
 
 const initialState = {
