@@ -2,7 +2,7 @@
  * Created by jiljiang on 2016/10/15.
  */
 
-import React, { PropTypes } from 'react';
+import React from 'react'; import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withLang } from 'containers/Lang';
 import cssModules from 'react-css-modules';
